chore: use consistent path
